requirements.txt
Flask==2.3.2
gunicorn==20.1.0
docker~=6.1.3
gitpython
PyJWT~=2.8.0
python-gitlab~=4.1.1
APScheduler~=3.10.4
flask-swagger-ui
flask-restful